Moves some files around
